requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
httpx==0.25.2
tenacity==8.2.3
vaderSentiment==3.3.2
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3
requests==2.31.0
moviepy==1.0.3
Pillow==10.1.0
python-dotenv==1.0.0

# Task 6 Dependencies
sentry-sdk==1.38.0
posthog==3.0.2
psycopg2-binary==2.9.9

# Jinja2 Templates
jinja2==3.1.2

# Database Migrations
alembic==1.13.1

# Testing
pytest==8.2.0
pytest-httpx==0.21.3
pytest-asyncio==0.23.0
requests==2.31.0